pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "cognitiveos"
version = "0.1.0"
description = "Local-first cognitive graph middleware with CLI, MCP, and Docker runtime targets."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
  { name = "Thermo Fisher Scientific / Bruce Feng" }
]
keywords = ["cognitive graph", "memory", "mcp", "cli", "sqlite"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: Other/Proprietary License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Database",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "beautifulsoup4>=4.12,<5.0",
  "httpx>=0.28,<1.0",
  "mcp[cli]>=1.27,<2.0",
  "pydantic>=2.11,<3.0",
  "python-dotenv>=1.0,<2.0",
  "sqlite-vec>=0.1.9,<0.2",
  "trafilatura>=2.0,<3.0",
  "typer>=0.16,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "ruff>=0.11,<0.12",
]
local-hf = [
  "sentence-transformers>=3.4,<4.0",
  "transformers>=4.48,<5.0",
]

[project.scripts]
cognitiveos = "cognitiveos.cli.app:main"
cognitiveos-mcp = "cognitiveos.mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cognitiveos"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
